Finish Basic Design

+++ TennisStoreClient/PrivateModels/Order.cs
@@ -4,10 +4,10 @@
     {
         public int Id { get; set; }
         public string? Name { get; set; }
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         public int Quantity { get; set; }
         public string? Image { get; set; }
-        public decimal SubTotal => Quantity * Price;
+        public double SubTotal => Quantity * Price;
         
     }
 }
